rename file after core type
