+++ dotcom-rendering/src/cricketMatch.ts
@@ -3,7 +3,7 @@ import type { FEFootballPageConfig } from './feFootballDataPage';
 import type {
 	FECricketInnings,
 	FECricketMatch,
-} from './frontend/feCricketMatch';
+} from './frontend/feCricketMatchPage';
 import type { EditionId } from './lib/edition';
 import { error, ok, type Result } from './lib/result';
 import type { NavType } from './model/extract-nav';
